requirements.txt
numpy
pandas
matplotlib
pillow
opencv-python
easyocr
pycocotools
requests
tqdm
streamlit
flask
